+++ mono_melos/packages/header_module/lib/header_widget.dart
@@ -1,7 +1,10 @@
 import 'package:flutter/material.dart';
 
 class HeaderWidget extends StatefulWidget {
-  const HeaderWidget({Key? key}) : super(key: key);
+  final Color headerColor;
+
+  const HeaderWidget(this.headerColor, {Key? key}) : super(key: key);
+
   @override
   State<HeaderWidget> createState() => _HeaderWidgetState();
 }
@@ -10,9 +13,9 @@ class _HeaderWidgetState extends State<HeaderWidget> {
   @override
   Widget build(BuildContext context) {
     return Container(
-      color: Colors.black,
-      width: 200,
-      height: 200,
+      color: widget.headerColor,
+      width: double.infinity,
+      height: 80,
     );
   }
 }