+++ mono_melos/packages/http_utility/lib/src/http_utility_if.dart
@@ -1,3 +1,3 @@
 abstract class HttpUtilityIF {
-  Map<String, String> getHeaders();
+  Map<String, String> getHeaders(String acceptHeaderValue);
 }