+++ mono_melos/packages/http_utility/lib/src/http_utility_impl.dart
@@ -2,7 +2,7 @@ import 'http_utility_if.dart';
 
 class HttpUtilityImpl implements HttpUtilityIF {
   @override
-  Map<String, String> getHeaders() {
-    return {'Content-Type': 'text/plain'};
+  Map<String, String> getHeaders(String acceptHeaderValue) {
+    return {'Content-Type': 'text/plain', 'Accept': acceptHeaderValue};
   }
 }