+++ mono_melos/packages/header_module/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.5
+
+ - **FEAT**: Pass color as "headerColor" param into widget.
+
 ## 0.0.4
 
  - **FEAT**: add black color.