+++ mono_melos/packages/http_utility/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+
+* Add Accept header to headers with user provided value
+
 ## 0.0.1
 
 * Provide http headers with Content-Type value